+++ src/components/DetailsCard.js
@@ -3,11 +3,19 @@ import PropTypes from 'prop-types';
 import ReactPlayer from 'react-player/youtube';
 import { Link } from 'react-router-dom';
 
-import { getInProgress, doneRecipes } from '../services/APIs/APIlocalStorage';
+import ShareIcon from './ShareIcon';
+import FavoriteIcon from './FavoriteIcon';
+import Card from './Card';
 
+import { getInProgress, doneRecipes } from '../services/APIs/APIlocalStorage';
+import {
+  sendToFavoriteStorage,
+  rmFromFavoriteStorage,
+  takeFavStorage,
+} from '../services/APIs/APIlocalStorage';
 import { FoodsContext } from '../contexts/FoodsContext';
 
-function StoreRecipe(id, ingredients, storedRecipes, type) {
+function StoreRecipe(id, ingredients, type) {
   const newStorage = {
     ...getInProgress(),
     [type === 'food' ? 'meals' : 'cocktails']: { ...getInProgress(type), [id]: ingredients },
@@ -16,17 +24,32 @@ function StoreRecipe(id, ingredients, storedRecipes, type) {
 }
 
 function DetailsCard({ eat, type }) {
-  const storedInprogress = getInProgress(type);
   const [, { setFoodInproggress }] = useContext(FoodsContext);
-  const { id, video, category, ingredients, instructions, isAlcoholic } = eat;
+  const { id, name, srcImage, video, category, ingredients, instructions, isAlcoholic } = eat;
 
   const startRecipe = useCallback(() => {
     setFoodInproggress(eat);
-    StoreRecipe(eat.id, eat.ingredients, storedInprogress, type);
+    StoreRecipe(eat.id, eat.ingredients, type);
   }, [eat, type]);
+
+  const handleFavoriteStorage = useCallback((isToSend) => {
+    if (isToSend) return sendToFavoriteStorage(eat, type);
+    return rmFromFavoriteStorage(id);
+  }, [type]);
 
   return (
     <div>
+      <Card
+        key={id}
+        name={name}
+        srcImage={srcImage}
+        testid={{ title: 'recipe-title', img: 'recipe-photo' }}
+      />
+      <ShareIcon textToCopy={window.location.href} />
+      <FavoriteIcon
+        handleFavoriteChange={handleFavoriteStorage}
+        isFavoriteInit={takeFavStorage().some((favorite) => Number(favorite.id) === Number(id))}
+      />
       <p data-testid="recipe-category">{isAlcoholic || category}</p>
       <ul>
         {ingredients.map(({ ingredient, measure }, index) => (
@@ -43,7 +66,7 @@ function DetailsCard({ eat, type }) {
             data-testid="start-recipe-btn"
             className="buttonIniciar"
             onClick={startRecipe}
-          >{storedInprogress[id] ? 'Continuar Receita' : 'Iniciar Receita'}</button>
+          >{getInProgress(type)[id] ? 'Continuar Receita' : 'Iniciar Receita'}</button>
         </Link>
       }
     </div>
